+++ src/libcore/mstream.cpp
@@ -23,10 +23,10 @@ void MemoryStream::read(void *p, size_t size) {
     if (m_pos + size > m_size) {
         const auto old_pos = m_pos;
         // Use signed difference since `m_pos` might be beyond `m_size`
-        int64_t sizeRead = m_size - m_pos;
+        int64_t sizeRead = m_size - static_cast<int64_t>(m_pos);
         if (sizeRead > 0) {
-            memcpy(p, m_data + m_pos, sizeRead);
-            m_pos += sizeRead;
+            memcpy(p, m_data + m_pos, static_cast<size_t>(sizeRead));
+            m_pos += static_cast<size_t>(sizeRead);
         }
         Log(EError, "Reading over the end of a memory stream!"
                     " (amount requested = %llu, amount actually read = %llu,"
